management-ui/mqtt_collector/requirements.txt
paho-mqtt
pymongo
flatten_json